use filepath.* to make goli crossplatform
also: do not change the current working directory: it changes semantics.
the user executes *goli* or other commands from the current working
directory. if *goli* now changes the current working directory the spawned
process inherits that change.

additionally: add -gopath flag to change the name of the marker file; it's
easier for folks to create a _gopath file on msft-windows than to create a
.gopath file.